backend/requirements.txt
crewai>=0.80.0
crewai-tools
fastapi
uvicorn
python-dotenv
pydantic
aiofiles